+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(hydrolook)
+
+test_check("hydrolook")